DS:update

+++ DSproject/pcmaker/src/pcmaker/Cooler.java
@@ -2,7 +2,7 @@ package pcmaker;
 
 public class Cooler extends pc_parts{
 
-	int as;//as기간
-	int air_vol;//풍량
-	int noisy;//소음
+	double as;//as기간
+	double air_vol;//풍량
+	double noisy;//소음
 }
