+++ DSproject/pcmaker/src/pcmaker/HDD.java
@@ -2,10 +2,9 @@ package pcmaker;
 
 public class HDD extends pc_parts{//SSD랑 나눈 이유는 SSD 읽고 쓰는 속도가 HDD보다 엄청나서 둘이 나눠놓음
 
-	int vol;//용량
-	int rspeed;//읽
-	int wspeed;//쓰 속도
-	int as;//as기간
+	double vol;//용량
+	double rspeed;//읽
+	double wspeed;//쓰 속도
+	double as;//as기간
 	
-
 }
