+++ DSproject/pcmaker/src/pcmaker/SSD.java
@@ -2,8 +2,8 @@ package pcmaker;
 
 public class SSD extends pc_parts{
 
-	int vol;//용량
-	int rspeed;//읽
-	int wspeed;//쓰 속도
-	int as;//as기간
+	double vol;//용량
+	double rspeed;//읽
+	double wspeed;//쓰 속도
+	double as;//as기간
 }
